Initial Commit
Genesis.py: can provide basic functionality to post request and receive responce. Moreover, because it is a class, so other python script can now use it without changing code in Genesis.
DescriptionProvider.py: is just an example on how to use Genesis.py. Its scenario is to send image to GenAI and get description from it.
Genesis.ipynb: just a lab file, all the experiment will be conducted there.