Final Update
I am so regret to anounce that this is the v0.1.0 and the LAST version of Genesis.py as OpenAI has forbidden Hong Kong user especially me from OpenRouter, and I do not have any other solutions but use the VPN like the others. There will be no longer update or bug correction for this version.